Back button support
Back button for movie detail to return to image list.  Not the best
implementation, but it will do for now.  Still some issues with the
menu, due largely to not understanding how to properly implement the
action bar/toolbar.

Also includes the images used for this level of the app implementation.